注入工具nlohmann json解析方式修改为cjson

+++ tools/event_inject/src/processing_finger_device.cpp
@@ -48,8 +48,8 @@ void ProcessingFingerDevice::AnalysisTouchPadFingerDate(const std::vector<Device
 {
     TouchPadCoordinates touchPadCoordinates = {};
     TouchPadInputEvent touchPadInputEvent = {};
-    for (uint32_t i = 0; i < inputData.size(); i++) {
-        for (uint32_t j = 0; j < inputData[i].posXY.size(); j++) {
+    for (size_t i = 0; i < inputData.size(); i++) {
+        for (size_t j = 0; j < inputData[i].posXY.size(); j++) {
             touchPadCoordinates.xPos = inputData[i].posXY[j].xPos;
             touchPadCoordinates.yPos = inputData[i].posXY[j].yPos;
             touchPadInputEvent.events.push_back(touchPadCoordinates);
